move utcdatetime to types module
